requirements.txt
fastapi==0.111.0
pydantic_settings==2.3.1
filetype==1.2.0
loguru==0.7.2
miniopy_async==1.19
